requirements.txt
swarms
tickr-agent
pydantic
loguru
swarm-models
fastapi 
uvicorn 
requests 
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0

black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pylint>=3.0.0
ruff>=0.1.0


mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0


ipython>=8.20.0
jupyter>=1.0.0
ipdb>=0.13.0
rich>=13.7.0

types-requests>=2.31.0
pandas-stubs>=2.0.0